requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
DateTime==4.3
freezefrog==0.3.2
future==0.17.1
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.11.0
py==1.8.0
pyArango==1.3.2
pycodestyle==2.5.0
pylint==2.3.1
pytest==4.5.0
pytz==2019.1
PyYAML==5.1
redis==2.10.6
requests==2.22.0
six==1.12.0
structlog==19.1.0
tasktiger==0.9.4
typed-ast==1.3.5
urllib3==1.25.2
wcwidth==0.1.7
wrapt==1.11.1
zope.interface==4.6.0
